requirements.txt
selenium
python-dotenv
openpyxl
pandas